practice/1016, 1017/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
Django==4.2.16
django-extensions==3.2.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.53.1
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
zipp==3.19.2
